로그인 성공 시 Home으로 전환, 화면 구성 및 controller 파일 분리
