+++ MuziGit/Controller/NotificationViewController.swift
@@ -1,0 +1,29 @@
+//
+//  NotificationViewController.swift
+//  MuziGit
+//
+//  Created by 이지은 on 8/7/24.
+//
+
+import UIKit
+
+class NotificationViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
